Update the Parser class
The updated parser will traverse the mjml tree and return array
consisting of mjml node objects with proper nesting structure.

This will enable us to traverse the nodes in order to rebuild the
HTML more easily.

Could probably add additional interfaces such as traversable, and
some of the itterator interfaces so that it's easier to traverse the
structure while rendering the HTML.